rte: move resourceobserver in its own file
Only trivial code movement, no intended changes in behaviour.
